Split out Aggregation and sort
