+++ src/Database/V5/Bloodhound/Types.hs
@@ -414,112 +414,16 @@ module Database.V5.Bloodhound.Types
 
 import           Bloodhound.Import
 
-import qualified Data.HashMap.Strict                   as HM
-import qualified Data.Map.Strict                       as M
-import qualified Data.Text                             as T
+import qualified Data.HashMap.Strict as HM
+import qualified Data.Map.Strict as M
 
 import           Database.V5.Bloodhound.Types.Class
 import           Database.V5.Bloodhound.Internal.Analysis
+import           Database.V5.Bloodhound.Internal.Aggregation
 import           Database.V5.Bloodhound.Internal.Client
 import           Database.V5.Bloodhound.Internal.Newtypes
 import           Database.V5.Bloodhound.Internal.Query
-
-{-| 'Sort' is a synonym for a list of 'SortSpec's. Sort behavior is order
-    dependent with later sorts acting as tie-breakers for earlier sorts.
--}
-type Sort = [SortSpec]
-
-{-| The two main kinds of 'SortSpec' are 'DefaultSortSpec' and
-    'GeoDistanceSortSpec'. The latter takes a 'SortOrder', 'GeoPoint', and
-    'DistanceUnit' to express "nearness" to a single geographical point as a
-    sort specification.
-
-<http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-request-sort.html#search-request-sort>
--}
-data SortSpec =
-    DefaultSortSpec DefaultSort
-  | GeoDistanceSortSpec SortOrder GeoPoint DistanceUnit
-  deriving (Eq, Show)
-
-instance ToJSON SortSpec where
-  toJSON (DefaultSortSpec
-          (DefaultSort (FieldName dsSortFieldName) dsSortOrder dsIgnoreUnmapped
-           dsSortMode dsMissingSort dsNestedFilter)) =
-    object [dsSortFieldName .= omitNulls base] where
-      base = [ "order" .= dsSortOrder
-             , "unmapped_type" .= dsIgnoreUnmapped
-             , "mode" .= dsSortMode
-             , "missing" .= dsMissingSort
-             , "nested_filter" .= dsNestedFilter ]
-
-  toJSON (GeoDistanceSortSpec gdsSortOrder (GeoPoint (FieldName field) gdsLatLon) units) =
-    object [ "unit" .= units
-           , field .= gdsLatLon
-           , "order" .= gdsSortOrder ]
-
-{-| 'DefaultSort' is usually the kind of 'SortSpec' you'll want. There's a
-    'mkSort' convenience function for when you want to specify only the most
-    common parameters.
-
-    The `ignoreUnmapped`, when `Just` field is used to set the elastic 'unmapped_type'
-
-<http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-request-sort.html#search-request-sort>
--}
-data DefaultSort =
-  DefaultSort { sortFieldName  :: FieldName
-              , sortOrder      :: SortOrder
-                                  -- default False
-              , ignoreUnmapped :: Maybe Text
-              , sortMode       :: Maybe SortMode
-              , missingSort    :: Maybe Missing
-              , nestedFilter   :: Maybe Filter } deriving (Eq, Show)
-
-{-| 'SortOrder' is 'Ascending' or 'Descending', as you might expect. These get
-    encoded into "asc" or "desc" when turned into JSON.
-
-<http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-request-sort.html#search-request-sort>
--}
-data SortOrder = Ascending
-               | Descending deriving (Eq, Show)
-
-instance ToJSON SortOrder where
-  toJSON Ascending  = String "asc"
-  toJSON Descending = String "desc"
-
-{-| 'Missing' prescribes how to handle missing fields. A missing field can be
-    sorted last, first, or using a custom value as a substitute.
-
-<http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-request-sort.html#_missing_values>
--}
-data Missing = LastMissing
-             | FirstMissing
-             | CustomMissing Text deriving (Eq, Show)
-
-instance ToJSON Missing where
-  toJSON LastMissing         = String "_last"
-  toJSON FirstMissing        = String "_first"
-  toJSON (CustomMissing txt) = String txt
-
-{-| 'SortMode' prescribes how to handle sorting array/multi-valued fields.
-
-http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-request-sort.html#_sort_mode_option
--}
-data SortMode = SortMin
-              | SortMax
-              | SortSum
-              | SortAvg deriving (Eq, Show)
-
-instance ToJSON SortMode where
-  toJSON SortMin = String "min"
-  toJSON SortMax = String "max"
-  toJSON SortSum = String "sum"
-  toJSON SortAvg = String "avg"
-
-{-| 'mkSort' defaults everything but the 'FieldName' and the 'SortOrder' so
-    that you can concisely describe the usual kind of 'SortSpec's you want.
--}
-mkSort :: FieldName -> SortOrder -> DefaultSort
-mkSort fieldName sOrder = DefaultSort fieldName sOrder Nothing Nothing Nothing Nothing
+import           Database.V5.Bloodhound.Internal.Sort
 
 {-| 'unpackId' is a silly convenience function that gets used once.
 -}
@@ -527,8 +431,6 @@ unpackId :: DocId -> Text
 unpackId (DocId docId) = docId
 
 type TrackSortScores = Bool
-newtype From = From Int deriving (Eq, Show, ToJSON)
-newtype Size = Size Int deriving (Eq, Show, ToJSON, FromJSON)
 
 data Search = Search { queryBody       :: Maybe Query
                      , filterBody      :: Maybe Filter
@@ -545,6 +447,31 @@ data Search = Search { queryBody       :: Maybe Query
                      , source          :: Maybe Source
                      , suggestBody     :: Maybe Suggest -- ^ Only one Suggestion request / response per Search is supported.
                      } deriving (Eq, Show)
+
+
+instance ToJSON Search where
+  toJSON (Search mquery sFilter sort searchAggs
+          highlight sTrackSortScores sFrom sSize _ sFields
+          sScriptFields sSource sSuggest) =
+    omitNulls [ "query"         .= query'
+              , "sort"          .= sort
+              , "aggregations"  .= searchAggs
+              , "highlight"     .= highlight
+              , "from"          .= sFrom
+              , "size"          .= sSize
+              , "track_scores"  .= sTrackSortScores
+              , "fields"        .= sFields
+              , "script_fields" .= sScriptFields
+              , "_source"       .= sSource
+              , "suggest"       .= sSuggest]
+
+    where query' = case sFilter of
+                    Nothing -> mquery
+                    Just x ->
+                        Just
+                      . QueryBoolQuery
+                      $ mkBoolQuery (maybeToList mquery)
+                        [x] [] []
 
 data SearchType = SearchTypeQueryThenFetch
                 | SearchTypeDfsQueryThenFetch
@@ -644,7 +571,6 @@ data SearchResult a =
                }
   deriving (Eq, Show)
 
-
 instance (FromJSON a) => FromJSON (SearchResult a) where
   parseJSON (Object v) = SearchResult <$>
                          v .:  "took"         <*>
@@ -674,7 +600,8 @@ instance (FromJSON a) => FromJSON (SearchHits a) where
   parseJSON _          = empty
 
 instance Semigroup (SearchHits a) where
-  (SearchHits ta ma ha) <> (SearchHits tb mb hb) = SearchHits (ta + tb) (max ma mb) (ha <> hb)
+  (SearchHits ta ma ha) <> (SearchHits tb mb hb) =
+    SearchHits (ta + tb) (max ma mb) (ha <> hb)
 
 instance Monoid (SearchHits a) where
   mempty = SearchHits 0 Nothing mempty
@@ -709,297 +636,6 @@ instance FromJSON HitFields where
     = HitFields <$> parseJSON x
 
 type HitHighlight = M.Map Text [Text]
-
-type Aggregations = M.Map Text Aggregation
-
-emptyAggregations :: Aggregations
-emptyAggregations = M.empty
-
-mkAggregations :: Text -> Aggregation -> Aggregations
-mkAggregations name aggregation = M.insert name aggregation emptyAggregations
-
-data TermOrder = TermOrder
-  { termSortField :: Text
-  , termSortOrder :: SortOrder } deriving (Eq, Show)
-
-instance ToJSON TermOrder where
-  toJSON (TermOrder termSortField termSortOrder) =
-    object [termSortField .= termSortOrder]
-
-data TermInclusion = TermInclusion Text
-                   | TermPattern Text Text deriving (Eq, Show)
-
-instance ToJSON TermInclusion where
-  toJSON (TermInclusion x) = toJSON x
-  toJSON (TermPattern pattern flags) =
-    omitNulls [ "pattern" .= pattern
-              , "flags"   .= flags]
-
-data CollectionMode = BreadthFirst
-                    | DepthFirst deriving (Eq, Show)
-
-instance ToJSON CollectionMode where
-  toJSON BreadthFirst = "breadth_first"
-  toJSON DepthFirst   = "depth_first"
-
-data ExecutionHint = Ordinals
-                   | GlobalOrdinals
-                   | GlobalOrdinalsHash
-                   | GlobalOrdinalsLowCardinality
-                   | Map deriving (Eq, Show)
-
-instance ToJSON ExecutionHint where
-  toJSON Ordinals                     = "ordinals"
-  toJSON GlobalOrdinals               = "global_ordinals"
-  toJSON GlobalOrdinalsHash           = "global_ordinals_hash"
-  toJSON GlobalOrdinalsLowCardinality = "global_ordinals_low_cardinality"
-  toJSON Map                          = "map"
-
-data Aggregation = TermsAgg TermsAggregation
-                 | CardinalityAgg CardinalityAggregation
-                 | DateHistogramAgg DateHistogramAggregation
-                 | ValueCountAgg ValueCountAggregation
-                 | FilterAgg FilterAggregation
-                 | DateRangeAgg DateRangeAggregation
-                 | MissingAgg MissingAggregation
-                 | TopHitsAgg TopHitsAggregation
-                 | StatsAgg StatisticsAggregation
-  deriving (Eq, Show)
-
-instance ToJSON Aggregation where
-  toJSON (TermsAgg (TermsAggregation term include exclude order minDocCount size shardSize collectMode executionHint termAggs)) =
-    omitNulls ["terms" .= omitNulls [ toJSON' term,
-                                      "include"        .= include,
-                                      "exclude"        .= exclude,
-                                      "order"          .= order,
-                                      "min_doc_count"  .= minDocCount,
-                                      "size"           .= size,
-                                      "shard_size"     .= shardSize,
-                                      "collect_mode"   .= collectMode,
-                                      "execution_hint" .= executionHint
-                                    ],
-               "aggs"  .= termAggs ]
-    where
-      toJSON' x = case x of { Left y -> "field" .= y;  Right y -> "script" .= y }
-
-  toJSON (CardinalityAgg (CardinalityAggregation field precisionThreshold)) =
-    object ["cardinality" .= omitNulls [ "field"              .= field,
-                                         "precisionThreshold" .= precisionThreshold
-                                       ]
-           ]
-
-  toJSON (DateHistogramAgg
-          (DateHistogramAggregation field interval format
-           preZone postZone preOffset postOffset dateHistoAggs)) =
-    omitNulls ["date_histogram" .= omitNulls [ "field"       .= field,
-                                               "interval"    .= interval,
-                                               "format"      .= format,
-                                               "pre_zone"    .= preZone,
-                                               "post_zone"   .= postZone,
-                                               "pre_offset"  .= preOffset,
-                                               "post_offset" .= postOffset
-                                             ],
-               "aggs"           .= dateHistoAggs ]
-  toJSON (ValueCountAgg a) = object ["value_count" .= v]
-    where v = case a of
-                (FieldValueCount (FieldName n)) ->
-                  object ["field" .= n]
-                (ScriptValueCount s) ->
-                  object ["script" .= s]
-  toJSON (FilterAgg (FilterAggregation filt ags)) =
-    omitNulls [ "filter" .= filt
-              , "aggs" .= ags]
-  toJSON (DateRangeAgg a) = object [ "date_range" .= a
-                                   ]
-  toJSON (MissingAgg (MissingAggregation{..})) =
-    object ["missing" .= object ["field" .= maField]]
-
-  toJSON (TopHitsAgg (TopHitsAggregation mfrom msize msort)) =
-    omitNulls ["top_hits" .= omitNulls [ "size" .= msize
-                                       , "from" .= mfrom
-                                       , "sort" .= msort
-                                       ]
-              ]
-
-  toJSON (StatsAgg (StatisticsAggregation typ field)) =
-    object [stType .= omitNulls [ "field" .= field ]]
-    where
-      stType | typ == Basic = "stats"
-             | otherwise = "extended_stats"
-
-data TopHitsAggregation = TopHitsAggregation
-  { taFrom :: Maybe From
-  , taSize :: Maybe Size
-  , taSort :: Maybe Sort
-  } deriving (Eq, Show)
-
-data MissingAggregation = MissingAggregation
-  { maField :: Text
-  } deriving (Eq, Show)
-
-data TermsAggregation = TermsAggregation
-  { term              :: Either Text Text
-  , termInclude       :: Maybe TermInclusion
-  , termExclude       :: Maybe TermInclusion
-  , termOrder         :: Maybe TermOrder
-  , termMinDocCount   :: Maybe Int
-  , termSize          :: Maybe Int
-  , termShardSize     :: Maybe Int
-  , termCollectMode   :: Maybe CollectionMode
-  , termExecutionHint :: Maybe ExecutionHint
-  , termAggs          :: Maybe Aggregations
-  } deriving (Eq, Show)
-
-data CardinalityAggregation = CardinalityAggregation
-  { cardinalityField   :: FieldName,
-    precisionThreshold :: Maybe Int
-  } deriving (Eq, Show)
-
-data DateHistogramAggregation = DateHistogramAggregation
-  { dateField      :: FieldName
-  , dateInterval   :: Interval
-  , dateFormat     :: Maybe Text
-    -- pre and post deprecated in 1.5
-  , datePreZone    :: Maybe Text
-  , datePostZone   :: Maybe Text
-  , datePreOffset  :: Maybe Text
-  , datePostOffset :: Maybe Text
-  , dateAggs       :: Maybe Aggregations
-  } deriving (Eq, Show)
-
-data DateRangeAggregation = DateRangeAggregation
-  { draField  :: FieldName
-  , draFormat :: Maybe Text
-  , draRanges :: NonEmpty DateRangeAggRange
-  } deriving (Eq, Show)
-
-instance ToJSON DateRangeAggregation where
-  toJSON DateRangeAggregation {..} =
-    omitNulls [ "field" .= draField
-              , "format" .= draFormat
-              , "ranges" .= toList draRanges
-              ]
-
-data DateRangeAggRange =
-    DateRangeFrom DateMathExpr
-  | DateRangeTo DateMathExpr
-  | DateRangeFromAndTo DateMathExpr DateMathExpr
-  deriving (Eq, Show)
-
-instance ToJSON DateRangeAggRange where
-  toJSON (DateRangeFrom e)        = object [ "from" .= e ]
-  toJSON (DateRangeTo e)          = object [ "to" .= e ]
-  toJSON (DateRangeFromAndTo f t) = object [ "from" .= f, "to" .= t ]
-
--- | See <https://www.elastic.co/guide/en/elasticsearch/reference/current/common-options.html#date-math> for more information.
-data DateMathExpr =
-  DateMathExpr DateMathAnchor [DateMathModifier]
-  deriving (Eq, Show)
-
-instance ToJSON DateMathExpr where
-  toJSON (DateMathExpr a mods) = String (fmtA a <> mconcat (fmtMod <$> mods))
-    where fmtA DMNow         = "now"
-          fmtA (DMDate date) = (T.pack $ showGregorian date) <> "||"
-          fmtMod (AddTime n u)      = "+" <> showText n <> fmtU u
-          fmtMod (SubtractTime n u) = "-" <> showText n <> fmtU u
-          fmtMod (RoundDownTo u)    = "/" <> fmtU u
-          fmtU DMYear   = "y"
-          fmtU DMMonth  = "M"
-          fmtU DMWeek   = "w"
-          fmtU DMDay    = "d"
-          fmtU DMHour   = "h"
-          fmtU DMMinute = "m"
-          fmtU DMSecond = "s"
-
--- | Starting point for a date range. This along with the 'DateMathModifiers' gets you the date ES will start from.
-data DateMathAnchor =
-    DMNow
-  | DMDate Day
-  deriving (Eq, Show)
-
-data DateMathModifier =
-    AddTime Int DateMathUnit
-  | SubtractTime Int DateMathUnit
-  | RoundDownTo DateMathUnit
-  deriving (Eq, Show)
-
-data DateMathUnit =
-    DMYear
-  | DMMonth
-  | DMWeek
-  | DMDay
-  | DMHour
-  | DMMinute
-  | DMSecond
-  deriving (Eq, Show)
-
--- | See <https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-valuecount-aggregation.html> for more information.
-data ValueCountAggregation =
-    FieldValueCount FieldName
-  | ScriptValueCount Script
-  deriving (Eq, Show)
-
--- | Single-bucket filter aggregations. See <https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-filter-aggregation.html#search-aggregations-bucket-filter-aggregation> for more information.
-data FilterAggregation = FilterAggregation
-  { faFilter :: Filter
-  , faAggs   :: Maybe Aggregations }
-  deriving (Eq, Show)
-
-data StatisticsAggregation = StatisticsAggregation
-  { statsType :: StatsType
-  , statsField :: FieldName }
-  deriving (Eq, Show)
-
-data StatsType
-  = Basic
-  | Extended
-  deriving (Eq, Show)
-
-mkTermsAggregation :: Text -> TermsAggregation
-mkTermsAggregation t =
-  TermsAggregation (Left t)
-  Nothing Nothing Nothing Nothing Nothing Nothing Nothing Nothing Nothing
-
-mkTermsScriptAggregation :: Text -> TermsAggregation
-mkTermsScriptAggregation t = TermsAggregation (Right t) Nothing Nothing Nothing Nothing Nothing Nothing Nothing Nothing Nothing
-
-mkDateHistogram :: FieldName -> Interval -> DateHistogramAggregation
-mkDateHistogram t i = DateHistogramAggregation t i Nothing Nothing Nothing Nothing Nothing Nothing
-
-mkCardinalityAggregation :: FieldName -> CardinalityAggregation
-mkCardinalityAggregation t = CardinalityAggregation t Nothing
-
-mkStatsAggregation :: FieldName -> StatisticsAggregation
-mkStatsAggregation = StatisticsAggregation Basic
-
-mkExtendedStatsAggregation :: FieldName -> StatisticsAggregation
-mkExtendedStatsAggregation = StatisticsAggregation Extended
-
-type AggregationResults = M.Map Text Value
-
-class BucketAggregation a where
-  key :: a -> BucketValue
-  docCount :: a -> Int
-  aggs :: a -> Maybe AggregationResults
-
-data Bucket a = Bucket
-  { buckets :: [a]
-  } deriving (Read, Show)
-
-instance (FromJSON a) => FromJSON (Bucket a) where
-  parseJSON (Object v) = Bucket <$>
-                         v .: "buckets"
-  parseJSON _ = mempty
-
-data BucketValue = TextValue Text
-                 | ScientificValue Scientific
-                 | BoolValue Bool deriving (Read, Show)
-
-instance FromJSON BucketValue where
-  parseJSON (String t) = return $ TextValue t
-  parseJSON (Number s) = return $ ScientificValue s
-  parseJSON (Bool b)   = return $ BoolValue b
-  parseJSON _          = mempty
 
 data MissingResult = MissingResult
   { missingDocCount :: Int
@@ -1093,7 +729,7 @@ instance BucketAggregation DateRangeResult where
   docCount = dateRangeDocCount
   aggs = dateRangeAggs
 
-toTerms :: Text -> AggregationResults ->  Maybe (Bucket TermsResult)
+toTerms :: Text -> AggregationResults -> Maybe (Bucket TermsResult)
 toTerms = toAggResult
 
 toDateHistogram :: Text -> AggregationResults -> Maybe (Bucket DateHistogramResult)
@@ -1117,30 +753,6 @@ getNamedSubAgg o knownKeys = maggRes
         maggRes
           | HM.null unknownKeys = Nothing
           | otherwise           = Just . M.fromList $ HM.toList unknownKeys
-
-instance ToJSON Search where
-  toJSON (Search mquery sFilter sort searchAggs
-          highlight sTrackSortScores sFrom sSize _ sFields
-          sScriptFields sSource sSuggest) =
-    omitNulls [ "query"         .= query'
-              , "sort"          .= sort
-              , "aggregations"  .= searchAggs
-              , "highlight"     .= highlight
-              , "from"          .= sFrom
-              , "size"          .= sSize
-              , "track_scores"  .= sTrackSortScores
-              , "fields"        .= sFields
-              , "script_fields" .= sScriptFields
-              , "_source"       .= sSource
-              , "suggest"       .= sSuggest]
-
-    where query' = case sFilter of
-                    Nothing -> mquery
-                    Just x ->
-                        Just
-                      . QueryBoolQuery
-                      $ mkBoolQuery (maybeToList mquery)
-                        [x] [] []
 
 instance ToJSON Source where
     toJSON NoSource                         = toJSON False
